+++ esw-backend-testkit/src/main/scala/esw/backend/testkit/stubs/SequenceManagerStubImpl.scala
@@ -75,7 +75,7 @@ class SequenceManagerStubImpl extends SequenceManagerApi {
       )
     )
 
-  override def getObsModesWithStatus: Future[ObsModesWithStatusResponse] = ???
+  override def getObsModesDetails: Future[ObsModesDetailsResponse] = ???
 }
 
 class SequenceManagerStub(val locationService: LocationService)(implicit val actorSystem: ActorSystem[SpawnProtocol.Command])