+++ esw-integration-test/src/test/scala/esw/integration/IntegrationTestWithAuth.scala
@@ -22,13 +22,13 @@ import esw.commons.utils.location.LocationServiceUtil
 import esw.gateway.api.clients.ClientFactory
 import esw.gateway.server.{GatewaySetup, GatewayWiring}
 import esw.ocs.api.actor.client.{SequenceComponentImpl, SequencerImpl}
-import esw.ocs.api.models.ObsModeStatus.{Configurable, Configured, NonConfigurable}
-import esw.ocs.api.models.{ObsMode, ObsModeWithStatus}
+import esw.sm.api.models.ObsModeStatus.{Configurable, Configured, NonConfigurable}
+import esw.ocs.api.models.ObsMode
 import esw.ocs.api.protocol.SequenceComponentResponse.SequencerLocation
 import esw.ocs.testkit.EswTestKit
 import esw.ocs.testkit.Service.AAS
 import esw.sm.api.models.ResourceStatus.{Available, InUse}
-import esw.sm.api.models.{AgentStatus, ProvisionConfig, Resource, SequenceComponentStatus}
+import esw.sm.api.models.{AgentStatus, ObsModeDetails, ProvisionConfig, Resource, Resources, SequenceComponentStatus}
 import esw.sm.api.protocol.CommonFailure.LocationServiceError
 import esw.sm.api.protocol.ConfigureResponse.{ConfigurationMissing, ConflictingResourcesWithRunningObsMode}
 import esw.sm.api.protocol.StartSequencerResponse.{LoadScriptError, SequenceComponentNotAvailable}
@@ -857,12 +857,17 @@ class IntegrationTestWithAuth extends EswTestKit(AAS) with GatewaySetup with Age
       TestSetup.cleanup()
     }
 
-    "getObsModesWithStatus should return all ObsModes with their status | ESW-466" in {
+    "getObsModesDetails should return all ObsModes with their status | ESW-466" in {
       locationService.unregisterAll().futureValue
       registerKeycloak()
       val eswSeqCompPrefix   = Prefix(ESW, "primary")
       val irisSeqCompPrefix  = Prefix(IRIS, "primary")
       val aoeswSeqCompPrefix = Prefix(AOESW, "primary")
+      val irisResource       = Resource(IRIS)
+      val nscuResource       = Resource(NSCU)
+      val nfiraosResource    = Resource(NFIRAOS)
+      val tcsResource        = Resource(TCS)
+      val wfosResource       = Resource(WFOS)
 
       TestSetup.startSequenceComponents(eswSeqCompPrefix, irisSeqCompPrefix, aoeswSeqCompPrefix)
 
@@ -870,15 +875,15 @@ class IntegrationTestWithAuth extends EswTestKit(AAS) with GatewaySetup with Age
 
       sequenceManager.configure(IRIS_CAL).futureValue
 
-      val expectedMessage = ObsModesWithStatusResponse.Success(
+      val expectedMessage = ObsModesDetailsResponse.Success(
         Set(
-          ObsModeWithStatus(IRIS_CAL, Configured),
-          ObsModeWithStatus(WFOS_CAL, Configurable),
-          ObsModeWithStatus(IRIS_DARKNIGHT, NonConfigurable)
+          ObsModeDetails(IRIS_CAL, Configured, Resources(irisResource, nfiraosResource, nscuResource)),
+          ObsModeDetails(WFOS_CAL, Configurable, Resources(wfosResource)),
+          ObsModeDetails(IRIS_DARKNIGHT, NonConfigurable, Resources(irisResource, tcsResource, nfiraosResource))
         )
       )
 
-      val actualResponse = sequenceManager.getObsModesWithStatus.futureValue.asInstanceOf[ObsModesWithStatusResponse.Success]
+      val actualResponse = sequenceManager.getObsModesDetails.futureValue.asInstanceOf[ObsModesDetailsResponse.Success]
       actualResponse should ===(expectedMessage)
 
       sequenceManager.shutdownAllSequenceComponents().futureValue