+++ esw-sm/esw-sm-api/shared/src/main/scala/esw/sm/api/protocol/SequenceManagerRequest.scala
@@ -9,7 +9,7 @@ sealed trait SequenceManagerRequest
 object SequenceManagerRequest {
   case class Configure(obsMode: ObsMode)        extends SequenceManagerRequest
   case class Provision(config: ProvisionConfig) extends SequenceManagerRequest
-  case object GetObsModesWithStatus             extends SequenceManagerRequest
+  case object GetObsModesDetails                extends SequenceManagerRequest
 
   case class StartSequencer(subsystem: Subsystem, obsMode: ObsMode)   extends SequenceManagerRequest
   case class RestartSequencer(subsystem: Subsystem, obsMode: ObsMode) extends SequenceManagerRequest